requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
openai==1.32.0
python-dotenv==1.0.1